requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
black==23.7.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.5
crashtest==0.4.1
distlib==0.3.6
dulwich==0.21.5
fastapi==0.100.0
filelock==3.12.2
flake8==6.0.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
iso8601==1.1.0
jaraco.classes==3.3.0
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
keyring==23.13.1
lockfile==0.12.2
mccabe==0.7.0
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.8.1
pluggy==1.2.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
ptyprocess==0.7.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.11
pydantic-settings==2.0.1
pydantic_core==2.1.2
pyflakes==3.0.1
pypika-tortoise==0.1.6
pyproject_hooks==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
pytz==2023.3
rapidfuzz==2.15.1
referencing==0.29.1
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.8.10
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy~=1.4.49
starlette==0.27.0
tomlkit==0.11.8
tortoise-orm==0.19.3
trove-classifiers==2023.7.6
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
virtualenv==20.23.1
webencodings==0.5.1
xattr==0.10.1
yarl==1.9.2
zipp==3.16.1

databases~=0.7.0
alembic~=1.11.1
APScheduler~=3.10.1